Move toml config parser to seperate package
